+++ rust/hyperlane-base/src/types/multisig.rs
@@ -5,19 +5,76 @@ use derive_new::new;
 use eyre::Result;
 use tracing::{debug, instrument};
 
-use hyperlane_core::{MultisigSignedCheckpoint, SignedCheckpointWithMessageId, H160, H256};
+use hyperlane_core::{
+    HyperlaneDomain, MultisigSignedCheckpoint, SignedCheckpointWithMessageId, H160, H256,
+};
 
-use crate::CheckpointSyncer;
+use crate::{CheckpointSyncer, CoreMetrics};
 
-/// Fetches signed checkpoints from multiple validators to create
-/// MultisigSignedCheckpoints
+/// For a particular validator set, fetches signed checkpoints from multiple
+/// validators to create MultisigSignedCheckpoints.
 #[derive(Clone, Debug, new)]
 pub struct MultisigCheckpointSyncer {
     /// The checkpoint syncer for each valid validator signer address
     checkpoint_syncers: HashMap<H160, Arc<dyn CheckpointSyncer>>,
+    metrics: Arc<CoreMetrics>,
+    app_context: Option<String>,
 }
 
 impl MultisigCheckpointSyncer {
+    /// Gets the latest checkpoint index from each validator's checkpoint syncer.
+    /// Returns a vector of the latest indices, in an unspecified order, and does
+    /// not contain indices for validators that did not provide a latest index.
+    /// Also updates the validator latest checkpoint metrics.
+    pub async fn get_validator_latest_checkpoints_and_update_metrics(
+        &self,
+        validators: &[H256],
+        origin: &HyperlaneDomain,
+        destination: &HyperlaneDomain,
+    ) -> Vec<u32> {
+        // Get the latest_index from each validator's checkpoint syncer.
+        // If a validator does not return a latest index, None is recorded so
+        // this can be surfaced in the metrics.
+        let mut latest_indices: HashMap<H160, Option<u32>> =
+            HashMap::with_capacity(validators.len());
+
+        for validator in validators {
+            let address = H160::from(*validator);
+            if let Some(checkpoint_syncer) = self.checkpoint_syncers.get(&address) {
+                // Gracefully handle errors getting the latest_index
+                match checkpoint_syncer.latest_index().await {
+                    Ok(Some(index)) => {
+                        debug!(?address, ?index, "Validator returned latest index");
+                        latest_indices.insert(H160::from(*validator), Some(index));
+                    }
+                    result => {
+                        debug!(
+                            ?address,
+                            ?result,
+                            "Failed to get latest index from validator"
+                        );
+                        latest_indices.insert(H160::from(*validator), None);
+                    }
+                }
+            }
+        }
+
+        if let Some(app_context) = &self.app_context {
+            self.metrics
+                .validator_metrics
+                .set_validator_latest_checkpoints(
+                    origin,
+                    destination,
+                    app_context.clone(),
+                    &latest_indices,
+                )
+                .await;
+        }
+
+        // Filter out any validators that did not return a latest index
+        latest_indices.values().copied().flatten().collect()
+    }
+
     /// Attempts to get the latest checkpoint with a quorum of signatures among
     /// validators.
     ///
@@ -37,24 +94,13 @@ impl MultisigCheckpointSyncer {
         threshold: usize,
         minimum_index: u32,
         maximum_index: u32,
+        origin: &HyperlaneDomain,
+        destination: &HyperlaneDomain,
     ) -> Result<Option<MultisigSignedCheckpoint>> {
-        // Get the latest_index from each validator's checkpoint syncer.
-        let mut latest_indices = Vec::with_capacity(validators.len());
-        for validator in validators {
-            let address = H160::from(*validator);
-            if let Some(checkpoint_syncer) = self.checkpoint_syncers.get(&address) {
-                // Gracefully handle errors getting the latest_index
-                match checkpoint_syncer.latest_index().await {
-                    Ok(Some(index)) => {
-                        debug!(?address, ?index, "Validator returned latest index");
-                        latest_indices.push(index);
-                    }
-                    err => {
-                        debug!(?address, ?err, "Failed to get latest index from validator");
-                    }
-                }
-            }
-        }
+        let mut latest_indices = self
+            .get_validator_latest_checkpoints_and_update_metrics(validators, origin, destination)
+            .await;
+
         debug!(
             ?latest_indices,
             "Fetched latest indices from checkpoint syncers"