+++ typescript/infra/src/config/chain.ts
@@ -2,64 +2,44 @@ import { providers } from 'ethers';
 
 import {
   ChainName,
-  RetryJsonRpcProvider,
-  RetryProviderOptions,
+  HyperlaneSmartProvider,
+  ProviderRetryOptions,
   RpcConsensusType,
+  chainMetadata,
 } from '@hyperlane-xyz/sdk';
 
 import { getSecretRpcEndpoint } from '../agents';
 
 import { DeployEnvironment } from './environment';
 
-export const defaultRetry = {
-  maxRequests: 6,
-  baseRetryMs: 50,
+export const defaultRetry: ProviderRetryOptions = {
+  maxRetries: 6,
+  baseRetryDelayMs: 50,
 };
-
-function buildProvider(config?: {
-  url?: string;
-  network?: providers.Networkish;
-  retry?: RetryProviderOptions;
-}): providers.JsonRpcProvider {
-  return config?.retry
-    ? new RetryJsonRpcProvider(config.retry, config?.url, config?.network)
-    : new providers.StaticJsonRpcProvider(config?.url, config?.network);
-}
 
 export async function fetchProvider(
   environment: DeployEnvironment,
   chainName: ChainName,
   connectionType: RpcConsensusType = RpcConsensusType.Single,
 ): Promise<providers.Provider> {
+  const chainId = chainMetadata[chainName].chainId;
   const single = connectionType === RpcConsensusType.Single;
   const rpcData = await getSecretRpcEndpoint(environment, chainName, !single);
-  switch (connectionType) {
-    case RpcConsensusType.Single: {
-      return buildProvider({ url: rpcData[0], retry: defaultRetry });
-    }
-    case RpcConsensusType.Quorum: {
-      return new providers.FallbackProvider(
-        (rpcData as string[]).map((url) => buildProvider({ url })), // disable retry for quorum
-      );
-    }
-    case RpcConsensusType.Fallback: {
-      return new providers.FallbackProvider(
-        (rpcData as string[]).map((url, index) => {
-          const fallbackProviderConfig: providers.FallbackProviderConfig = {
-            provider: buildProvider({ url, retry: defaultRetry }),
-            // Priority is used by the FallbackProvider to determine
-            // how to order providers using ascending ordering.
-            // When not specified, all providers have the same priority
-            // and are ordered randomly for each RPC.
-            priority: index,
-          };
-          return fallbackProviderConfig;
-        }),
-        1, // a single provider is "quorum", but failure will cause failover to the next provider
-      );
-    }
-    default: {
-      throw Error(`Unsupported connectionType: ${connectionType}`);
-    }
+
+  if (connectionType === RpcConsensusType.Single) {
+    return HyperlaneSmartProvider.fromRpcUrl(chainId, rpcData[0], defaultRetry);
+  } else if (
+    connectionType === RpcConsensusType.Quorum ||
+    connectionType === RpcConsensusType.Fallback
+  ) {
+    return new HyperlaneSmartProvider(
+      chainId,
+      rpcData.map((url) => ({ http: url })),
+      undefined,
+      // disable retry for quorum
+      connectionType === RpcConsensusType.Fallback ? defaultRetry : undefined,
+    );
+  } else {
+    throw Error(`Unsupported connectionType: ${connectionType}`);
   }
 }