+++ typescript/sdk/src/consts/multisigIsm.ts
@@ -18,6 +18,7 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0xec68258a7c882ac2fc46b81ce80380054ffb4ef2', // dsrv
       '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
       '0x092e1c19da58e87ea65198785ee83867fe4bb418', // everstake
+      '0xc2d68e109a7e80e12098d50ac4ef9fa7b3061684', // staked
     ],
   },
 
@@ -43,8 +44,10 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     threshold: 2,
     validators: [
       '0xb9453d675e0fa3c178a17b4ce1ad5b1a279b3af9',
-      '0x4512985a574cb127b2af2d4bb676876ce804e3f8',
+      '0x64889519ad3ffc8f3ae436fcd63efc6b853fd63f', // staked
       '0x41188cb5a5493a961c467ba38a3f8b1f1d35ee63', // everstake
+      '0xcff391b4e516452d424db66beb9052b041a9ed79', // dsrv
+      '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
     ],
   },
 
@@ -61,8 +64,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     threshold: 2,
     validators: [
       '0x570af9b7b36568c8877eebba6c6727aa9dab7268',
-      '0x8292b1a53907ece0f76af8a50724e9492bcdc8a3', // bsc
+      '0x8292b1a53907ece0f76af8a50724e9492bcdc8a3', // dsrv
       '0xeaf5cf9100f36a4baeea779f8745dda86159103c', // everstake
+      '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
     ],
   },
 
@@ -81,6 +85,8 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x63478422679303c3e4fc611b771fa4a707ef7f4a',
       '0x622e43baf06ad808ca8399360d9a2d9a1a12688b', // dsrv
       '0xf2c1e3888eb618f1f1a071ef3618f134715a9a49', // everstake
+      '0x46ecbc794574727abb8f97f01dacd9db6135f47a', // staked
+      '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
     ],
   },
 
@@ -100,6 +106,7 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x94438a7de38d4548ae54df5c6010c4ebc5239eae', // dsrv
       '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
       '0xce327111035dd38698c92c3778884dbbb0ca8103', // everstake
+      '0xb2f5a6a6e6046e2ede213617e989329666a6c4bc', // staked
     ],
   },
 
@@ -118,6 +125,7 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0xd4df66a859585678f2ea8357161d896be19cc1ca',
       '0x19fb7e04a1be6b39b6966a0b0c60b929a93ed672', // dsrv
       '0xdb96116d13a2fadde9742d7cc88474a5ed39a03a', // everstake
+      '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
     ],
   },
 
@@ -167,6 +175,7 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x2225e2f4e9221049456da93b71d2de41f3b6b2a8',
       '0x645428d198d2e76cbd9c1647f5c80740bb750b97', // dsrv
       '0xaed886392df07897743d8e272d438f00c4c9a2ae', // everstake
+      '0xcf0bb43255849cb3709a96ee166e5c3ce4adc7f9', // staked
     ],
   },
 
@@ -187,14 +196,6 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x7885fae56dbcf5176657f54adbbd881dc6714132',
     ],
   },
-  neutrontestnet: {
-    threshold: 2,
-    validators: [
-      '0x5d2a99d67cd294a821de4fb25da6901ea8f89814',
-      '0xb57486243ce3bb3c38c50a582b8bbd20cb393589',
-      '0x661faee997654d14ead4ae48035883f05c3150cf',
-    ],
-  },
 
   optimism: {
     threshold: 2,
@@ -202,6 +203,8 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
       '0x20349eadc6c72e94ce38268b96692b1a5c20de4f',
       '0x5b7d47b76c69740462432f6a5a0ca5005e014157', // dsrv
       '0x22b1ad4322cdb5f2c76ebf4e5a93803d480fcf0d', // everstake
+      '0x9636fbe90b6816438327b0fbde435aa3c8eeda15', // staked
+      '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
     ],
   },
 
@@ -218,8 +221,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     threshold: 2,
     validators: [
       '0x12ecb319c7f4e8ac5eb5226662aeb8528c5cefac',
-      '0x8dd8f8d34b5ecaa5f66de24b01acd7b8461c3916',
+      '0x008f24cbb1cc30ad0f19f2516ca75730e37efb5f', // dsrv
       '0x722aa4d45387009684582bca8281440d16b8b40f', // everstake
+      '0x5450447aee7b544c462c9352bef7cad049b0c2dc', // zeeprime
     ],
   },
 
@@ -227,7 +231,7 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     threshold: 2,
     validators: [
       '0x86f2a44592bb98da766e880cfd70d3bbb295e61a',
-      '0xc84076030bdabaabb9e61161d833dd84b700afda',
+      '0x865818fe1db986036d5fd0466dcd462562436d1a', // dsrv
       '0x57231619fea13d85270ca6943298046c75a6dd01', // everstake
     ],
   },
@@ -245,8 +249,9 @@ export const defaultMultisigConfigs: ChainMap<MultisigConfig> = {
     threshold: 2,
     validators: [
       '0xad557170a9f2f21c35e03de07cb30dcbcc3dff63',
-      '0xb37fe43a9f47b7024c2d5ae22526cc66b5261533',
+      '0x37148DE77D9FA915e6F0A9B54bCdF5e6f53ca511', // staked
       '0x276de8e2b88e659c4e5ad30d62d9de42c3da3403', // everstake
+      '0xbac4ac39f1d8b5ef15f26fdb1294a7c9aba3f948', // dsrv
     ],
   },