+++ typescript/sdk/src/core/HyperlaneCore.ts
@@ -1,4 +1,5 @@
 import { ethers } from 'ethers';
+import type { TransactionReceipt as ViemTxReceipt } from 'viem';
 
 import { Mailbox__factory } from '@hyperlane-xyz/core';
 import {
@@ -119,7 +120,7 @@ export class HyperlaneCore extends HyperlaneApp<CoreFactories> {
   }
 
   waitForMessageProcessing(
-    sourceTx: ethers.ContractReceipt,
+    sourceTx: ethers.ContractReceipt | ViemTxReceipt,
   ): Promise<ethers.ContractReceipt[]> {
     const messages = HyperlaneCore.getDispatchedMessages(sourceTx);
     return Promise.all(messages.map((msg) => this.waitForProcessReceipt(msg)));
@@ -127,7 +128,7 @@ export class HyperlaneCore extends HyperlaneApp<CoreFactories> {
 
   // TODO consider renaming this, all the waitForMessage* methods are confusing
   async waitForMessageProcessed(
-    sourceTx: ethers.ContractReceipt,
+    sourceTx: ethers.ContractReceipt | ViemTxReceipt,
     delay?: number,
     maxAttempts?: number,
   ): Promise<void> {
@@ -146,12 +147,14 @@ export class HyperlaneCore extends HyperlaneApp<CoreFactories> {
   }
 
   // Redundant with static method but keeping for backwards compatibility
-  getDispatchedMessages(sourceTx: ethers.ContractReceipt): DispatchedMessage[] {
+  getDispatchedMessages(
+    sourceTx: ethers.ContractReceipt | ViemTxReceipt,
+  ): DispatchedMessage[] {
     return HyperlaneCore.getDispatchedMessages(sourceTx);
   }
 
   static getDispatchedMessages(
-    sourceTx: ethers.ContractReceipt,
+    sourceTx: ethers.ContractReceipt | ViemTxReceipt,
   ): DispatchedMessage[] {
     const mailbox = Mailbox__factory.createInterface();
     const dispatchLogs = sourceTx.logs