+++ typescript/utils/index.ts
@@ -18,6 +18,7 @@ export {
   eqAddressEvm,
   eqAddressSol,
   getAddressProtocolType,
+  isAddress,
   isAddressCosmos,
   isAddressEvm,
   isAddressSealevel,
@@ -48,6 +49,7 @@ export {
 export { chunk, exclude } from './src/arrays';
 export {
   pollAsync,
+  raceWithContext,
   retryAsync,
   runWithTimeout,
   sleep,
@@ -83,6 +85,7 @@ export {
 export {
   ValueOf,
   arrayToObject,
+  deepCopy,
   deepEquals,
   invertKeysAndValues,
   isObject,