+++ typescript/cli/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @hyperlane-xyz/cli
 
+## 3.6.0
+
+### Patch Changes
+
+- 67a6d971e: Added `shouldRecover` flag to deployContractFromFactory so that the `TestRecipientDeployer` can deploy new contracts if it's not the owner of the prior deployments (We were recovering the SDK artifacts which meant the deployer won't be able to set the ISM as they needed)
+- Updated dependencies [67a6d971e]
+- Updated dependencies [612d4163a]
+- Updated dependencies [0488ef31d]
+- Updated dependencies [8d8ba3f7a]
+  - @hyperlane-xyz/sdk@3.6.0
+  - @hyperlane-xyz/utils@3.6.0
+
+## 3.5.1
+
+### Patch Changes
+
+- Updated dependencies [a04454d6d]
+  - @hyperlane-xyz/sdk@3.5.1
+  - @hyperlane-xyz/utils@3.5.1
+
+## 3.5.0
+
+### Patch Changes
+
+- 05a943b4a: Skip mandatory balance check for remotes in send commands"
+- Updated dependencies [655b6a0cd]
+- Updated dependencies [08ba0d32b]
+- Updated dependencies [f7d285e3a]
+  - @hyperlane-xyz/sdk@3.5.0
+  - @hyperlane-xyz/utils@3.5.0
+
+## 3.4.0
+
+### Patch Changes
+
+- e06fe0b32: Supporting DefaultFallbackRoutingIsm through non-factory deployments
+- dcf8b800a: Fixes for commands with --yes flag
+- 9c7dbcb94: Remove domainId and protocolType setting when creating chain config
+- Updated dependencies [7919417ec]
+- Updated dependencies [fd4fc1898]
+- Updated dependencies [e06fe0b32]
+- Updated dependencies [b832e57ae]
+- Updated dependencies [79c96d718]
+  - @hyperlane-xyz/sdk@3.4.0
+  - @hyperlane-xyz/utils@3.4.0
+
 ## 3.3.0
 
 ### Minor Changes