+++ typescript/sdk/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @hyperlane-xyz/sdk
 
+## 3.6.0
+
+### Minor Changes
+
+- 0488ef31d: Add dsrv, staked and zeeprime as validators
+- 8d8ba3f7a: HyperlaneIsmFactory is now wary of (try)getDomainId or (try)getChainName calls which may fail and handles them appropriately.
+
+### Patch Changes
+
+- 67a6d971e: Added `shouldRecover` flag to deployContractFromFactory so that the `TestRecipientDeployer` can deploy new contracts if it's not the owner of the prior deployments (We were recovering the SDK artifacts which meant the deployer won't be able to set the ISM as they needed)
+- 612d4163a: Add mailbox version const to SDK
+  - @hyperlane-xyz/core@3.6.0
+  - @hyperlane-xyz/utils@3.6.0
+
+## 3.5.1
+
+### Patch Changes
+
+- a04454d6d: Use getBalance instead of queryContractSmart for CwTokenAdapter
+  - @hyperlane-xyz/core@3.5.1
+  - @hyperlane-xyz/utils@3.5.1
+
+## 3.5.0
+
+### Minor Changes
+
+- 655b6a0cd: Redeploy Routing ISM Factories
+
+### Patch Changes
+
+- 08ba0d32b: Remove dead arbitrum goerli explorer link"
+- f7d285e3a: Adds Test Recipient addresses to the SDK artifacts
+  - @hyperlane-xyz/core@3.5.0
+  - @hyperlane-xyz/utils@3.5.0
+
+## 3.4.0
+
+### Minor Changes
+
+- b832e57ae: Replace Fallback and Retry Providers with new SmartProvider with more effective fallback/retry logic
+
+### Patch Changes
+
+- 7919417ec: Granular control of updating predeployed routingIsms based on routing config mismatch
+  - Add support for routingIsmDelta which filters out the incompatibility between the onchain deployed config and the desired config.
+  - Based on the above, you either update the deployed Ism with new routes, delete old routes, change owners, etc.
+  - `moduleMatchesConfig` uses the same
+- fd4fc1898: - Upgrade Viem to 1.20.0
+  - Add optional restUrls field to ChainMetadata
+  - Add deepCopy util function
+  - Add support for cosmos factory token addresses
+- e06fe0b32: Supporting DefaultFallbackRoutingIsm through non-factory deployments
+- 79c96d718: Remove healthy RPC URLs and remove NeutronTestnet
+- Updated dependencies [fd4fc1898]
+- Updated dependencies [e06fe0b32]
+  - @hyperlane-xyz/utils@3.4.0
+  - @hyperlane-xyz/core@3.4.0
+
 ## 3.3.0
 
 ### Patch Changes